+++ src/types/routerTypes.tsx
@@ -63,6 +63,7 @@ import type { SweepPrivateKeyProcessingParams } from '../components/scenes/Sweep
 import type { SweepPrivateKeySelectCryptoParams } from '../components/scenes/SweepPrivateKeySelectCryptoScene'
 import type { TransactionDetailsParams } from '../components/scenes/TransactionDetailsScene'
 import type { TransactionListParams } from '../components/scenes/TransactionListScene'
+import type { TransactionList2Params } from '../components/scenes/TransactionListScene2'
 import type { TransactionsExportParams } from '../components/scenes/TransactionsExportScene'
 import type { WcConnectionsParams } from '../components/scenes/WcConnectionsScene'
 import type { WcConnectParams } from '../components/scenes/WcConnectScene'
@@ -88,6 +89,7 @@ import type { FiatPluginSepaFormParams } from '../plugins/gui/scenes/SepaFormSce
 export type WalletsTabParamList = {} & {
   walletList: undefined
   transactionList: TransactionListParams
+  transactionList2: TransactionList2Params
   transactionDetails: TransactionDetailsParams
 }
 
@@ -204,6 +206,7 @@ export type EdgeAppStackParamList = {} & {
   sweepPrivateKeySelectCrypto: SweepPrivateKeySelectCryptoParams
   testScene: undefined
   transactionDetails: TransactionDetailsParams
+  transactionList2: TransactionList2Params
   transactionsExport: TransactionsExportParams
   upgradeUsername: undefined
   walletRestore: undefined